refactor(life): refactor component to a stateless form
